Rewrite codegen, implement subtraction, improve parser
Rewrites the codegen to collect every variable first, and then allocate
them at the beginning of the program. This way we can default to zero
when a variable is used which has not been manually initialized.
Implements subtraction correctly (does not allow negative numbers). This
is done using saturation arithmetic without branching. Also improves the
parser to report an error if a semicolon is missing.